Add query hook for any API method (#53)
## What does this PR do?

When no command is found, Kourou will try to execute the given command with the `sdk:query` command.  

The first argument has to be the name of the controller and the action separated by a semicolon (eg `document:create`)

Kourou will try to infer common arguments like `index`, `collection`, `_id` or `body`.  

It will automatically infer and accept the following lists of arguments:
 - `<command> <index>`
    * _eg: `kourou collection:list iot`_
.


 - `<command> <index> <collection>`
    * _eg: `kourou collection:truncate iot sensors`_
.


 - `<command> <index> <collection> <body>`
    * _eg: `kourou bulk:import iot sensors '{bulkData: []}'`_
.


 - `<command> <index> <collection> <id>`
    * _eg: `kourou document:delete iot sensors sigfox-123`_
.


 - `<command> <index> <collection> <id> <body>`
    * _eg: `kourou document:create iot sensors sigfox-123 '{temperature: 42}'`_

Then any argument will be passed as-is to the `sdk:query` method.


### Other changes

  - add `--id` argument to `sdk:query`
 - remove `document:get`
 - remove `document:create`